Split settings into multiple files
