+++ crates/utils/src/settings/mod.rs
@@ -1,153 +1,13 @@
-use crate::{location_info, LemmyError};
+use crate::{location_info, settings::structs::Settings, LemmyError};
 use anyhow::Context;
 use deser_hjson::from_str;
 use merge::Merge;
-use serde::Deserialize;
-use std::{
-  env,
-  fs,
-  io::Error,
-  net::{IpAddr, Ipv4Addr},
-  sync::RwLock,
-};
+use std::{env, fs, io::Error, sync::RwLock};
+
+pub mod defaults;
+pub mod structs;
 
 static CONFIG_FILE: &str = "config/config.hjson";
-
-#[derive(Debug, Deserialize, Clone, Merge)]
-pub struct Settings {
-  pub setup: Option<Setup>,
-  pub database: Option<DatabaseConfig>,
-  pub hostname: Option<String>,
-  pub bind: Option<IpAddr>,
-  pub port: Option<u16>,
-  pub tls_enabled: Option<bool>,
-  pub jwt_secret: Option<String>,
-  pub pictrs_url: Option<String>,
-  pub iframely_url: Option<String>,
-  pub rate_limit: Option<RateLimitConfig>,
-  pub email: Option<EmailConfig>,
-  pub federation: Option<FederationConfig>,
-  pub captcha: Option<CaptchaConfig>,
-}
-
-impl Default for Settings {
-  fn default() -> Self {
-    Self {
-      database: Some(DatabaseConfig::default()),
-      rate_limit: Some(RateLimitConfig::default()),
-      federation: Some(FederationConfig::default()),
-      captcha: Some(CaptchaConfig::default()),
-      email: None,
-      setup: None,
-      hostname: None,
-      bind: Some(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
-      port: Some(8536),
-      tls_enabled: Some(true),
-      jwt_secret: Some("changeme".into()),
-      pictrs_url: Some("http://pictrs:8080".into()),
-      iframely_url: Some("http://iframely".into()),
-    }
-  }
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct Setup {
-  pub admin_username: String,
-  pub admin_password: String,
-  pub admin_email: Option<String>,
-  pub site_name: String,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct RateLimitConfig {
-  pub message: i32,
-  pub message_per_second: i32,
-  pub post: i32,
-  pub post_per_second: i32,
-  pub register: i32,
-  pub register_per_second: i32,
-  pub image: i32,
-  pub image_per_second: i32,
-}
-
-impl Default for RateLimitConfig {
-  fn default() -> Self {
-    Self {
-      message: 180,
-      message_per_second: 60,
-      post: 6,
-      post_per_second: 600,
-      register: 3,
-      register_per_second: 3600,
-      image: 6,
-      image_per_second: 3600,
-    }
-  }
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct EmailConfig {
-  pub smtp_server: String,
-  pub smtp_login: Option<String>,
-  pub smtp_password: Option<String>,
-  pub smtp_from_address: String,
-  pub use_tls: bool,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct CaptchaConfig {
-  pub enabled: bool,
-  pub difficulty: String,
-}
-
-impl Default for CaptchaConfig {
-  fn default() -> Self {
-    Self {
-      enabled: true,
-      difficulty: "medium".into(),
-    }
-  }
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct DatabaseConfig {
-  pub user: String,
-  pub password: String,
-  pub host: String,
-  pub port: i32,
-  pub database: String,
-  pub pool_size: u32,
-}
-
-impl Default for DatabaseConfig {
-  fn default() -> Self {
-    Self {
-      user: "lemmy".into(),
-      password: "password".into(),
-      host: "localhost".into(),
-      port: 5432,
-      database: "lemmy".into(),
-      pool_size: 5,
-    }
-  }
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct FederationConfig {
-  pub enabled: bool,
-  pub allowed_instances: Option<String>,
-  pub blocked_instances: Option<String>,
-}
-
-impl Default for FederationConfig {
-  fn default() -> Self {
-    Self {
-      enabled: false,
-      allowed_instances: Some("".into()),
-      blocked_instances: Some("".into()),
-    }
-  }
-}
 
 lazy_static! {
   static ref SETTINGS: RwLock<Settings> = RwLock::new(match Settings::init() {
